finition admin

+++ src/Controller/UserController.php
@@ -25,7 +25,7 @@ use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 class UserController extends AbstractController
 {
     #[Route('/', name: 'app_user_index', methods: ['GET'])]
-    public function index(UserRepository $userRepository,TokenInterface $token, MailerInterface $mailer): Response
+    public function index(UserRepository $userRepository, TokenInterface $token, MailerInterface $mailer): Response
     {
         $user = $token->getUser();
 
@@ -33,7 +33,7 @@ class UserController extends AbstractController
     }
 
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager,TokenInterface $token, UserPasswordHasherInterface $hasher): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, TokenInterface $token, UserPasswordHasherInterface $hasher): Response
     {
         $user = $token->getUser();
 
@@ -48,7 +48,7 @@ class UserController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, User $user, EntityManagerInterface $entityManager, UserPasswordHasherInterface $passwordUser, TokenInterface $token,  Security $security): Response
+    public function edit(Request $request, User $user, EntityManagerInterface $entityManager, UserPasswordHasherInterface $passwordUser, TokenInterface $token, Security $security): Response
     {
         if ((!$this->isGranted('user_edit', $user) && $security->getUser() !== $user)) {
             $user = $token->getUser();
@@ -102,50 +102,44 @@ class UserController extends AbstractController
     }
     // #[Route('/{id}', name: 'app_user_delete', methods: ['POST']), isGranted("ROLE_ADMIN")]
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
-    public function delete(Request $request, User $user, EntityManagerInterface $entityManager, TokenInterface $token,  Security $security): Response
+    public function delete(Request $request, User $user, EntityManagerInterface $entityManager, TokenInterface $token, Security $security): Response
     {
         if ((!$this->isGranted('user_edit', $user) && $security->getUser() !== $user) || $this->isGranted('ROLE_ADMIN')) {
             $user = $token->getUser();
             return $this->redirectToRoute('app_user_edit', ["id" => $user->getProfil()->getId()], Response::HTTP_SEE_OTHER);
         }
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
-            $profil = $user->getProfil();
-    
-            // Supprimer les enregistrements liés à band_member_role
-            foreach ($profil->getBandMembers() as $member) {
-                foreach ($member->getBandMemberRoles() as $role) {
-                    $entityManager->remove($role);
+            $user->getProfil() ? $profil = $user->getProfil() : $profil = null;
+
+            if ($profil) {
+                foreach ($profil->getBandMembers() as $member) {
+                    foreach ($member->getBandMemberRoles() as $role) {
+                        $entityManager->remove($role);
+                    }
+                    $entityManager->remove($member);
                 }
-                $entityManager->remove($member);
+
+                foreach ($profil->getNotifications() as $notif) {
+                    $entityManager->remove($notif);
+                }
+
+                foreach ($profil->getHallMembers() as $member) {
+                    foreach ($member->getHallMemberRoles() as $role) {
+                        $entityManager->remove($role);
+                    }
+                    $entityManager->remove($member);
+                }
+                $entityManager->remove($profil);
             }
 
-            foreach ($profil->getNotifications() as $notif)
-            {
-                $entityManager->remove($notif);
-
-            }
-    
-            // Supprimer les enregistrements liés à hall_member_role
-            foreach ($profil->getHallMembers() as $member) {
-                foreach ($member->getHallMemberRoles() as $role) {
-                    $entityManager->remove($role);
-                }
-                $entityManager->remove($member);
-            }
-    
-            // Supprimer le profil
-            $entityManager->remove($profil);
-    
-            // Supprimer l'utilisateur
             $entityManager->remove($user);
-    
             $entityManager->flush();
             $this->addFlash(
-               'success',
-               'Votre profil à bien été supprimé, Bonne continuation !'
+                'success',
+                'Votre profil à bien été supprimé, Bonne continuation !'
             );
         }
-    
+
         return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
     }
 }
